Removed the decorator pattern from RestConnection
